# takeout-server/API-Swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "NodeJS Server for Take-Out App.",
    "version": "1.0.0",
    "title": "Take Out Server"
  },
  "host": "take-out-server.herokuapp.com",
  "schemes": [
    "https"
  ],
...
  "paths": {
...
    "/orders": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get all orders",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          },
          "400": {
            "description": "Failed"
          }
        }
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Creates new order",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userid": {
                  "type": "string"
                },
                "chefid": {
                  "type": "string"
                },
                "discountid": {
                  "type": "string"
                },
                "dishs": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "price": {
                  "type": "number"
                },
                "note": {
                  "type": "string"
                },
                "location": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": "string"
                    },
                    "longitude": {
                      "type": "number"
                    },
                    "latitude": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "401": {
            "description": "Failed"
          }
        }
      }
    },
...
    "/admin/login": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "For admin to login to app",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          },
          "401": {
            "description": "Unanthorized"
          }
        }
      }
    },
    "/chefs/all": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get chefs for administration",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Chef"
              }
            }
          },
          "400": {
            "description": "Failed"
          }
        }
      }
    },
    "/chefs/{id}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Activated/deactivated specified chef",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "activated": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Chef"
            }
          },
          "401": {
            "description": "Failed"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get all users",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "401": {
            "description": "Failed"
          }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Activated/Deactivated specified user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "activated": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Failed"
          }
        }
      }
    },
    "/admin/summary": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get admin summary",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Failed"
          }
        }
      }
    }
  },
  "definitions": {
    "Account": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "activated": {
          "type": "boolean"
        }
      }
    },
    "Admin": {
      "type": "object",
      "properties": {
        "account": {
          "$ref": "#/definitions/Account"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "account": {
          "$ref": "#/definitions/Account"
        },
        "name": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "homeLocation": {
          "type": "object",
          "properties": {
            "address": {
              "type": "string"
            },
            "longitude": {
              "type": "number"
            },
            "latitude": {
              "type": "number"
            }
          }
        },
        "workLocation": {
          "type": "object",
          "properties": {
            "address": {
              "type": "string"
            },
            "longitude": {
              "type": "number"
            },
            "latitude": {
              "type": "number"
            }
          }
        },
        "points": {
          "type": "number"
        }
      }
    },
    "Chef": {
      "type": "object",
      "properties": {
        "account": {
          "$ref": "#/definitions/Account"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "rating": {
          "type": "object",
          "properties": {
            "stars": {
              "type": "number"
            },
            "votes": {
              "type": "number"
            }
          }
        },
        "location": {
          "type": "object",
          "properties": {
            "address": {
              "type": "string"
            },
            "longitude": {
              "type": "number"
            },
            "latitude": {
              "type": "number"
            }
          }
        },
        "storehours": {
          "type": "object",
          "properties": {
            "open": {
              "type": "number"
            },
            "close": {
              "type": "number"
            },
            "saturday": {
              "type": "boolean"
            },
            "sunday": {
              "type": "boolean"
            }
          }
        },
        "profilePhoto": {
          "type": "string"
        },
        "licencePhoto": {
          "type": "string"
        },
        "cuisineType": {
          "type": "string"
        },
        "menus": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Menu"
          }
        },
        "paymentRatio": {
          "type": "number"
        }
      }
    },
    "Menu": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "dishs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Dish"
          }
        }
      }
    },
    "Dish": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "photo": {
          "type": "string"
        },
        "cuisineType": {
          "type": "string"
        },
        "cooktime": {
          "type": "number"
        }
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/User"
        },
        "chef": {
          "$ref": "#/definitions/Chef"
        },
        "dishs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Dish"
          }
        },
        "discount": {
          "$ref": "#/definitions/Discount"
        },
        "note": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "status": {
          "type": "string"
        },
        "accepted": {
          "type": "boolean"
        },
        "deliveryLocation": {
          "type": "object",
          "properties": {
            "address": {
              "type": "string"
            },
            "longitude": {
              "type": "number"
            },
            "latitude": {
              "type": "number"
            }
          }
        }
      }
    },
    "Discount": {
      "type": "object",
      "properties": {
        "codestring": {
          "type": "string"
        },
        "discountRate": {
          "type": "number"
        },
        "beenUsed": {
          "type": "boolean"
        }
      }
    }
  }
}